fix: address review feedback on cluster bootstrap retry
- collapse the duplicated Unreleased changelog heading
- demote per-attempt bootstrap failure logs to warn; exhaustion still
  surfaces as error from the actor system start path
- document that permanent errors also burn the retry budget, and how
  bootstrapMaxAttempts relates to Olric's MaxJoinAttempts (nested, not
  parallel)
- move the retry tests into cluster_test.go and cut the e2e test from
  ~47s to ~3s via the NATS provider's connect-retry knobs
